Restructuring the Rejuvenate Funcitonality a Bit
Changed the rejuvenate() function to expect a card or list of cards rather than a number of cards.  rejuvenateOne() and rejuvenateMany() now take a number and select the corresponding cards to pass to the rejuvenate() function.

Secondary changes, the notify call for rejuvenating will now be a single line listing all cards rejuvenated,rather than using a single line per card.  Also accounted Statdmiller's change in placement of the notification.  Also updated the rejuvenate() call in the automation to be rejuvenateOne().