Fixing a Small Bug
OCTGN seems to have a bug where if you use pile.top() or pile.bottom() on an empty pile, it will return card object with garbage content instead of either error out or returning a null value.

Checking the pile length to ensure there's a target to rejuvenate is good practice anyway.  It's not necessary in rejuvenateMany() due to that function's structure though.