Aligning Set Files
Change to some set files in a previous merge never made it to the live server so the recent commit bringing the repo up to date reverted them.  This commit is to reintroduce them.  Not the proper way to handle conflicts with Git but hopefully it won't cause too many problems.