Partial Merge Fix
Turns out the set files from another branch did not have cards added to some of the same set files creating an awkward situation where we want to accept changes to the same files from two sources.  Not sure how to properly address this.

As a first step in reconciliation, this commit should accept those changes that contain no conflicts.